requirements.txt
streamlit 
requests 
langchain 
transformers
python-dotenv
